+++ CHANGELOG.md
@@ -17,7 +17,7 @@ TBD
 
 Add support for the new CloudFront cache tags associated by WordPress post types.
 
-- Added new "Clear Cache by Tag" submenu option to clear the site cache by post type as cache tags
+- Added new "Clear Cache by Post Type" submenu option to clear the site cache by post type as cache tags
 - Added new header `x-amz-meta-cache-tag` for post type pages
 - Special pages not post type specific also have tags: 'error' (for error and 404 pages), 'home', 'search', 'archive', and 'author'.
 - Unknown pages use the tag 'unknown'